Breaking out filters.js into separate files
